refactor: Refactor api documantation rules
