+++ src/ed_domain/documentation/api/abc_api_client.py
@@ -1,0 +1,14 @@
+from abc import ABCMeta, abstractmethod
+from typing import Generic, TypeVar
+
+from ed_domain.documentation.api.definitions import (ApiResponse,
+                                                     EndpointCallParams)
+
+ResponseType = TypeVar("ResponseType")
+
+
+class ABCApiClient(Generic[ResponseType], metaclass=ABCMeta):
+    @abstractmethod
+    def __call__(
+        self, call_params: EndpointCallParams
+    ) -> ApiResponse[ResponseType]: ...
