+++ src/ed_domain/documentation/api/definitions/__init__.py
@@ -1,0 +1,13 @@
+from ed_domain.documentation.api.definitions.api_response import ApiResponse
+from ed_domain.documentation.api.definitions.endpoint_call_params import \
+    EndpointCallParams
+from ed_domain.documentation.api.definitions.endpoint_description import \
+    EndpointDescription
+from ed_domain.documentation.api.definitions.http_method import HttpMethod
+
+__all__ = [
+    "HttpMethod",
+    "ApiResponse",
+    "EndpointCallParams",
+    "EndpointDescription",
+]
