+++ src/ed_domain/documentation/api/definitions/endpoint_description.py
@@ -1,6 +1,6 @@
 from typing import NotRequired, TypedDict
 
-from ed_domain.documentation.common.http_method import HttpMethod
+from ed_domain.documentation.api.definitions.http_method import HttpMethod
 
 
 class EndpointDescription(TypedDict):
